+++ app/src/main/java/com/example/klarnacodingchallenge/domain/WeatherInfoDataDaily.kt
@@ -9,7 +9,7 @@ import kotlinx.parcelize.Parcelize
  */
 @Parcelize
 data class WeatherInfoDataDaily(
-    val time: Int = Int.MIN_VALUE,
+    val time: Long = Long.MIN_VALUE,
     val summary: String = EMPTY,
     val icon: String = EMPTY,
     val sunriseTime: Int = Int.MIN_VALUE,
